Add useState for signed in users, moved sign in/out logic to another config file
