+++ src/App.js
@@ -5,31 +5,41 @@ import BottomInputComponent from "./components/BottomInputComponent"
 import RightSideComponent from "./components/RightSideComponent"
 import SignIn from "./components/SignIn"
 import { Routes, Route } from "react-router-dom"
+import { useState } from "react"
 
 function App() {
+  const [authState, setAuthState] = useState(null)
+
+  const handleAuthStateChange = (auth) => {
+    setAuthState(auth)
+  }
+
   return (
-    <Routes>
-      <Route
-        path='/signin'
-        element={
-          <>
-            <SignIn />
-          </>
-        }
-      />
-      <Route
-        path='/chat'
-        element={
-          <>
-            <TopNavigationBar />
-            <LeftSideComponent />
-            <MiddleChatWindow />
-            <BottomInputComponent />
-            <RightSideComponent />
-          </>
-        }
-      />
-    </Routes>
+    <>
+      <h5>{authState ? authState.email : "no user signed in"}</h5>
+      <Routes>
+        <Route
+          path='/'
+          element={
+            <>
+              <SignIn handleAuthStateChange={handleAuthStateChange} />
+            </>
+          }
+        />
+        <Route
+          path='/chat'
+          element={
+            <>
+              <TopNavigationBar />
+              <LeftSideComponent />
+              <MiddleChatWindow />
+              <BottomInputComponent />
+              <RightSideComponent />
+            </>
+          }
+        />
+      </Routes>
+    </>
   )
 }
