+++ src/components/SignIn.js
@@ -1,9 +1,16 @@
-import React from "react"
-import { signInWithGoogle } from "../config/fire-config"
+import { useEffect } from "react"
+import { signInWithGoogle, auth } from "../config/googleSignIn"
+import { onAuthStateChanged } from "firebase/auth"
 import logo from "../assets/sparrow-logo.svg"
 import { Link } from "react-router-dom"
 
-function SignIn() {
+function SignIn({ handleAuthStateChange }) {
+  useEffect(() => {
+    onAuthStateChanged(auth, (data) => {
+      handleAuthStateChange(data)
+    })
+  }, [])
+
   return (
     <div>
       <img
