+++ src/config/fire-config.js
@@ -2,21 +2,7 @@
 import { initializeApp } from "firebase/app"
 // TODO: Add SDKs for Firebase products that you want to use
 // https://firebase.google.com/docs/web/setup#available-libraries
-import {
-  getFirestore,
-  addDoc,
-  getDocs,
-  collection,
-  serverTimestamp,
-  query,
-  where,
-} from "firebase/firestore"
-import {
-  getAuth,
-  GoogleAuthProvider,
-  signInWithPopup,
-  signOut,
-} from "firebase/auth"
+import { getFirestore } from "firebase/firestore"
 
 // Your web app's Firebase configuration
 const firebaseConfig = {
@@ -30,42 +16,4 @@ const firebaseConfig = {
 
 // Initialize Firebase
 const app = initializeApp(firebaseConfig)
-// auth object is the gateway to the Firebase authentication API
-// -> we can reference auth objects to manage user accounts and credentials
-export const auth = getAuth(app)
 export const db = getFirestore(app)
-
-// provider object represents everything related to Google authentication
-const provider = new GoogleAuthProvider()
-const usersCollectionRef = collection(db, "users")
-
-// Prompt user to sign in w/their Google account by opening a pop-up window
-export const signInWithGoogle = () => {
-  signInWithPopup(auth, provider)
-    .then(async (result) => {
-      // Checks if user is already on "users" collection
-      const userDocs = await getDocs(
-        query(usersCollectionRef, where("email", "==", result.user.email))
-      )
-      const user = userDocs.docs.map((doc) => doc.data())
-      if (!user.length) {
-        addDoc(usersCollectionRef, {
-          name: result.user.displayName,
-          email: result.user.email,
-          photoURL: result.user.photoURL,
-          timestamp: serverTimestamp(),
-        })
-      }
-      // Information about the user based on who signed in
-      console.log(result.user)
-    })
-    .catch((error) => {
-      console.log(error)
-    })
-}
-
-// Signs-out of Friendly Chat.
-export const signOutUser = () => {
-  // Sign out of Firebase.
-  signOut(getAuth())
-}
